+++ E16/e/src/ewins.h
@@ -1,6 +1,6 @@
 /*
  * Copyright (C) 2000-2007 Carsten Haitzler, Geoff Harrison and various contributors
- * Copyright (C) 2004-2011 Kim Woelders
+ * Copyright (C) 2004-2012 Kim Woelders
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to
@@ -75,7 +75,7 @@ struct _ewin {
    struct {
       Win                 win;
       int                 x, y, w, h, bw;
-      long                event_mask;
+      unsigned int        event_mask;
    } client;
 
    struct {
