+++ E16/e/src/setup.c
@@ -1,6 +1,6 @@
 /*
  * Copyright (C) 2000-2007 Carsten Haitzler, Geoff Harrison and various contributors
- * Copyright (C) 2004-2010 Kim Woelders
+ * Copyright (C) 2004-2012 Kim Woelders
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to
@@ -46,7 +46,7 @@ SetupX(const char *dstr)
 {
    int                 err;
    char                buf[128];
-   long                mask;
+   unsigned int        mask;
 
    if (!dstr)
       dstr = getenv("DISPLAY");
